+++ docs/demo/js/index.js
@@ -1,20 +1,25 @@
 // The autoComplete.js Engine instance creator
 const autoCompleteJS = new autoComplete({
   data: {
-    src: async function () {
-      // Loading placeholder text
-      document.querySelector("#autoComplete").setAttribute("placeholder", "Loading...");
-      // Fetch External Data Source
-      const source = await fetch("../demo/db/generic.json");
-      const data = await source.json();
-      // Post Loading placeholder text
-      document.querySelector("#autoComplete").setAttribute("placeholder", autoCompleteJS.placeHolder);
-      // Returns Fetched data
-      return data;
+    src: async () => {
+      try {
+        // Loading placeholder text
+        autoCompleteJS.input.setAttribute("placeholder", "Loading...");
+        // Fetch External Data Source
+        const source = await fetch("./db/generic.json");
+        const data = await source.json();
+        // Post Loading placeholder text
+        autoCompleteJS.input.setAttribute("placeholder", autoCompleteJS.placeHolder);
+        // Returns Fetched data
+        return data;
+      } catch (error) {
+        return error;
+      }
     },
-    key: ["food", "cities", "animals"],
-    results: (list) => {
+    keys: ["food", "cities", "animals"],
+    filter: (list) => {
       // Filter duplicates
+      // incase of multiple data keys usage
       const filteredResults = Array.from(new Set(list.map((value) => value.match))).map((food) => {
         return list.find((value) => value.match === food);
       });
@@ -22,53 +27,85 @@ const autoCompleteJS = new autoComplete({
       return filteredResults;
     },
   },
-  trigger: {
-    event: ["input", "focus"],
-  },
   placeHolder: "Search for Food & Drinks!",
-  searchEngine: "strict",
   resultsList: {
-    noResults: (list, query) => {
-      // No Results List Message
-      const message = document.createElement("div");
-      message.setAttribute("class", "no_result");
-      message.innerHTML = `<span style="display: flex; align-items: center; font-weight: 100; color: rgba(0,0,0,.2);">Found No Results for "${query}"</span>`;
-      list.appendChild(message);
+    element: (list, data) => {
+      const info = document.createElement("p");
+      if (data.results.length > 0) {
+        info.innerHTML = `Displaying <strong>${data.results.length}</strong> out of <strong>${data.matches.length}</strong> results`;
+      } else {
+        info.innerHTML = `Found <strong>${data.matches.length}</strong> matching results for <strong>"${data.query}"</strong`;
+      }
+      list.prepend(info);
     },
+    noResults: true,
+    maxResults: 15,
+    tabSelect: true,
   },
   resultItem: {
-    content: (data, element) => {
+    element: (item, data) => {
       // Modify Results Item Style
-      element.style = "display: flex; justify-content: space-between;";
+      item.style = "display: flex; justify-content: space-between;";
       // Modify Results Item Content
-      element.innerHTML = `
-        <span style="text-overflow: ellipsis; white-space: nowrap; overflow: hidden;">
-            ${data.match}
-        </span>
-        <span style="display: flex; align-items: center; font-size: 13px; font-weight: 100; text-transform: uppercase; color: rgba(0,0,0,.2);">
-            ${data.key}
-        </span>`;
+      item.innerHTML = `
+      <span style="text-overflow: ellipsis; white-space: nowrap; overflow: hidden;">
+        ${data.match}
+      </span>
+      <span style="display: flex; align-items: center; font-size: 13px; font-weight: 100; text-transform: uppercase; color: rgba(0,0,0,.2);">
+        ${data.key}
+      </span>`;
     },
-    highlight: {
-      render: true,
-    },
+    highlight: true,
   },
-  onSelection: (feedback) => {
-    document.querySelector("#autoComplete").blur();
-    // Prepare User's Selected Value
-    const selection = feedback.selection.value[feedback.selection.key];
-    // Render selected choice to selection div
-    document.querySelector(".selection").innerHTML = selection;
-    // Replace Input value with the selected value
-    document.querySelector("#autoComplete").value = selection;
-    // Console log autoComplete data feedback
-    console.log(feedback);
+  events: {
+    input: {
+      focus: () => {
+        if (autoCompleteJS.input.value.length) autoCompleteJS.start();
+      },
+    },
   },
 });
 
+// autoCompleteJS.input.addEventListener("init", function (event) {
+//   console.log(event);
+// });
+
+// autoCompleteJS.input.addEventListener("response", function (event) {
+//   console.log(event.detail);
+// });
+
+// autoCompleteJS.input.addEventListener("results", function (event) {
+//   console.log(event.detail);
+// });
+
+// autoCompleteJS.input.addEventListener("open", function (event) {
+//   console.log(event.detail);
+// });
+
+// autoCompleteJS.input.addEventListener("navigate", function (event) {
+//   console.log(event.detail);
+// });
+
+autoCompleteJS.input.addEventListener("selection", function (event) {
+  const feedback = event.detail;
+  autoCompleteJS.input.blur();
+  // Prepare User's Selected Value
+  const selection = feedback.selection.value[feedback.selection.key];
+  // Render selected choice to selection div
+  document.querySelector(".selection").innerHTML = selection;
+  // Replace Input value with the selected value
+  autoCompleteJS.input.value = selection;
+  // Console log autoComplete data feedback
+  console.log(feedback);
+});
+
+// autoCompleteJS.input.addEventListener("close", function (event) {
+//   console.log(event.detail);
+// });
+
 // Toggle Search Engine Type/Mode
-document.querySelector(".toggler").addEventListener("click", function () {
-  // Holds the toggle button alignment
+document.querySelector(".toggler").addEventListener("click", () => {
+  // Holds the toggle button selection/alignment
   const toggle = document.querySelector(".toggle").style.justifyContent;
 
   if (toggle === "flex-start" || toggle === "") {
@@ -84,8 +121,8 @@ document.querySelector(".toggler").addEventListener("click", function () {
   }
 });
 
-// Toggle results list and other elements
-const action = function (action) {
+// Blur/unBlur page elements
+const action = (action) => {
   const github = document.querySelector(".github-corner");
   const title = document.querySelector("h1");
   const mode = document.querySelector(".mode");
@@ -107,15 +144,14 @@ const action = function (action) {
   }
 };
 
-// Toggle event for search input
-// showing & hiding results list onfocus/blur
-["focus", "blur"].forEach(function (eventType) {
-  document.querySelector("#autoComplete").addEventListener(eventType, function () {
-    // Hide results list & show other elements
+// Blur/unBlur page elements on input focus
+["focus", "blur"].forEach((eventType) => {
+  autoCompleteJS.input.addEventListener(eventType, () => {
+    // Blur page elements
     if (eventType === "blur") {
       action("dim");
     } else if (eventType === "focus") {
-      // Show results list & hide other elements
+      // unBlur page elements
       action("light");
     }
   });