+++ src/helpers/eventEmitter.js
@@ -1,0 +1,10 @@
+/**
+ * Event emitter/dispatcher
+ *
+ * @param {String} name - Name of fired event
+ * @param {Object} ctx - autoComplete.js context
+ */
+export default (name, ctx) => {
+  // Dispatch event on "inputField"
+  ctx.input.dispatchEvent(new CustomEvent(name, { bubbles: true, detail: ctx.feedback, cancelable: true }));
+};