+++ src/helpers/io.js
@@ -1,0 +1,120 @@
+/**
+ * DOM Element selector
+ *
+ * @param {String|HTMLElement} element - html tag | html element
+ *
+ * @returns {HTMLElement} - selected html element
+ */
+const select = (element) => (typeof element === "string" ? document.querySelector(element) : element);
+
+/**
+ * Create new element
+ *
+ * @param {String|HTMLElement} tag - html tag | html element
+ * @param {Object} options - of the html element
+ *
+ * @returns {HTMLElement} - created html element
+ */
+const create = (tag, options) => {
+  const el = typeof tag === "string" ? document.createElement(tag) : tag;
+
+  for (const key in options) {
+    const val = options[key];
+
+    if (key === "inside") {
+      val.append(el);
+    } else if (key === "dest") {
+      select(val[0]).insertAdjacentElement(val[1], el);
+    } else if (key === "around") {
+      const ref = select(val);
+
+      ref.parentNode.insertBefore(el, ref);
+
+      el.append(ref);
+
+      if (ref.getAttribute("autofocus") != null) ref.focus();
+    } else if (key in el) {
+      el[key] = val;
+    } else {
+      el.setAttribute(key, val);
+    }
+  }
+
+  return el;
+};
+
+/**
+ * Get the "inputField" query value
+ *
+ * @param {Element} field - input or textarea element
+ *
+ * @returns {String} - Raw query value as a string
+ */
+const getQuery = (field) =>
+  field instanceof HTMLInputElement || field instanceof HTMLTextAreaElement ? field.value : field.innerHTML;
+
+/**
+ * Format input value
+ *
+ * @param {String} value - user's raw search query value
+ * @param {Object} diacritics - formatting on/off
+ *
+ * @returns {String} - Raw "inputField" value as a string
+ */
+const format = (value, diacritics) => {
+  value = value.toLowerCase();
+
+  return (
+    diacritics
+      ? value
+          .normalize("NFD")
+          .replace(/[\u0300-\u036f]/g, "")
+          .normalize("NFC")
+      : value
+  ).toString();
+};
+
+/**
+ * Debouncer
+ *
+ * @param {Function} callback - Callback function
+ * @param {Number} duration - Delay time value
+ *
+ * @returns {Function} - Debouncer function
+ */
+const debounce = (callback, duration) => {
+  let timer;
+
+  return () => {
+    clearTimeout(timer);
+
+    timer = setTimeout(() => callback(), duration);
+  };
+};
+
+/**
+ * Trigger condition validator
+ *
+ * @param {String} query - User's manipulated search query value
+ * @param {Function} condition - trigger condition rule
+ * @param {Number} threshold - of query length to trigger
+ *
+ * @returns {Boolean} - For autoComplete.js to run or not
+ */
+const checkTrigger = (query, condition, threshold) => (condition ? condition(query) : query.length >= threshold);
+
+/**
+ * Highlight matching characters
+ *
+ * @param {String} value - user's raw search query value
+ * @param {Array} classes - of highlighted character
+ *
+ * @returns {HTMLElement} - newly create html element
+ */
+const mark = (value, classes) =>
+  create("mark", {
+    innerHTML: value,
+    ...(typeof classes === "string" && { classes }),
+  }).outerHTML;
+
+export { select, create, getQuery, format, debounce, checkTrigger, mark };